Fixed several type-conversion warnings passing gemm descriptor's batch-size (c1 or c2) into code generation only taking a 32-bit value of the respective count. Perform (0 - x) rather than -x in order to avoid warning about unary minus operator applied to unsigned type (result still unsigned). Avoid warnings about invariant conditions (/*const*/). Hoisted libxsmm_dnn_typesize in outer scope to avoid warnings about type-conversion involving size_t into 32-bit arithmetic.

+++ src/generator_gemm_amx_microkernel.c
@@ -871,7 +871,7 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
             n_CL_to_pf,
             i_xgemm_desc->lda * 2,
             i_gp_reg_mapping->gp_reg_a,
-            _A_offsets[i] + pf_dist * i_xgemm_desc->c1);
+            (unsigned int)(_A_offsets[i] + pf_dist * i_xgemm_desc->c1));
       }
     }
 
@@ -892,7 +892,7 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
             n_CL_to_pf,
             i_xgemm_desc->ldb * 2,
             i_gp_reg_mapping->gp_reg_b,
-            _B_offsets[i] + pf_dist * i_xgemm_desc->c2);
+            (unsigned int)(_B_offsets[i] + pf_dist * i_xgemm_desc->c2));
       }
     }
 
@@ -957,7 +957,7 @@ void libxsmm_generator_gemm_amx_kernel_kloop( libxsmm_generated_code*           
   LIBXSMM_UNUSED( io_loop_label_tracker );
 
   if (i_xgemm_desc->flags & LIBXSMM_GEMM_FLAG_BATCH_REDUCE_STRIDE) {
-    i_brgemm_loop = A_offs/i_xgemm_desc->c1;
+    i_brgemm_loop = A_offs/((unsigned int)i_xgemm_desc->c1);
   }
 
   while (i_xgemm_desc->k % l_k_blocking != 0) {
